pyproject.toml
[tool.poetry]
name = "lamparao"
version = "0.1.0"
description = ""
authors = ["cabrumcos"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pre-commit = "^3.5.0"
selenium = "^4.14.0"
prefect = "^2.13.8"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
sqlalchemy-libsql = "^0.1.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
